test: Move test helpers to __tests__/helpers/
Previously I updated the jest config that `uptrend-scripts` provides to
a custom match pattern to only match files with .test. in the name. This
was to allow helper modules to be created next in the __tests__ folders
to be used by the actual test files. Setting that custom config worked
but when I tried to merge my PR it failed because the code coverage was
way to low due to the fact jest considered the helper file not tested.

Looking at the `uptrend-scripts` default jest config I realized that
`__tests__/helpers/` was defined to be ignored completely when looking
for tests and what files to create covereage reports for.
It does this by setting them in jest config option
`testPathIgnorePatterns` and `coveragePathIgnorePatterns`.

So the real solution was to add a `helpers/` folder in the `__tests__`
dir I created a helper in.
